# New BatchSchema class to generate patch selectors and combine into batch selectors