--- a/xbatcher/generators.py
+++ b/xbatcher/generators.py
@@ -6,40 +6,43 @@ from typing import Any, Dict, Hashable, Iterator, List, OrderedDict, Sequence, U
 import xarray as xr
 
 
-def _gen_slices(dimsize: int, size: int, overlap: int = 0) -> Any:
+def _gen_slices(*, dim_size: int, slice_size: int, overlap: int = 0) -> Any:
     # return a list of slices to chop up a single dimension
-    if overlap >= size:
+    if overlap >= slice_size:
         raise ValueError(
             "input overlap must be less than the input sample length, but "
-            f"the input sample length is {size} and the overlap is {overlap}"
+            f"the input sample length is {slice_size} and the overlap is {overlap}"
         )
     slices = []
-    stride = size - overlap
-    for start in range(0, dimsize, stride):
-        end = start + size
-        if end <= dimsize:
+    stride = slice_size - overlap
+    for start in range(0, dim_size, stride):
+        end = start + slice_size
+        if end <= dim_size:
             slices.append(slice(start, end))
     return slices
 
 
 def _iterate_through_dataset(
     ds: Union[xr.Dataset, xr.DataArray],
+    *,
     dims: OrderedDict[Hashable, int],
     overlap: Dict[Hashable, int] = {},
 ) -> Any:
     dim_slices = []
     for dim in dims:
-        dimsize = ds.sizes[dim]
-        size = dims[dim]
+        dim_size = ds.sizes[dim]
+        slice_size = dims[dim]
         olap = overlap.get(dim, 0)
-        if size > dimsize:
+        if slice_size > dim_size:
             raise ValueError(
                 "input sample length must be less than or equal to the "
-                f"dimension length, but the sample length of {size} "
-                f"is greater than the dimension length of {dimsize} "
+                f"dimension length, but the sample length of {slice_size} "
+                f"is greater than the dimension length of {dim_size} "
                 f"for {dim}"
             )
-        dim_slices.append(_gen_slices(dimsize, size, olap))
+        dim_slices.append(
+            _gen_slices(dim_size=dim_size, slice_size=slice_size, overlap=olap)
+        )
 
     for slices in itertools.product(*dim_slices):
         selector = dict(zip(dims, slices))
@@ -191,7 +194,9 @@ class BatchGenerator:
         return dict(enumerate(batches))
 
     def _iterate_batch_dims(self) -> Any:
-        return _iterate_through_dataset(self.ds, self.batch_dims)
+        return _iterate_through_dataset(self.ds, dims=self.batch_dims)
 
     def _iterate_input_dims(self) -> Any:
-        return _iterate_through_dataset(self.ds, self.input_dims, self.input_overlap)
+        return _iterate_through_dataset(
+            self.ds, dims=self.input_dims, overlap=self.input_overlap
+        )