@andersy005 2022-12-20
@maxrjones, this looks great overall! i really like the graphics in https://github.com/xarray-contrib/xbatcher/pull/132#issue-1473258637. Having some of these graphics included in the docs would be awesome!